+++ Client/src/routes/routeConfig.jsx
@@ -2,7 +2,8 @@ import {createBrowserRouter} from 'react-router-dom';
 import Home from "../pages/HomePage.jsx";
 import {MainLayout} from "../layouts/MainLayout.jsx";
 import {PageNotFound} from "../pages/PageNotFound.jsx";
-import SignIn from '../modules/core/components/account/SingIn.jsx';
+import SignIn from '../modules/core/components/account/SignIn.jsx';
+
 
 export const router = createBrowserRouter([
   {
